Continued working on categories

+++ app/app/src/main/java/com/example/seng440assignment2/model/Beer.kt
@@ -17,5 +17,5 @@ data class BeerListItem(
     val title: String,
     val barcode: String,
     val photoURL: String,
-    val averageRating: Long
+    val averageRating: Float
 )
